feat(app): BackButton component
