+++ app/src/common/components/BackButton.tsx
@@ -1,0 +1,30 @@
+import React from "react";
+
+import { Button, ButtonProps, Typography } from "@material-ui/core";
+import ArrowBack from "@material-ui/icons/ArrowBackIos";
+import { useTranslation } from "react-i18next";
+import { useHistory } from "react-router-dom";
+
+type SourceEditButtonProps = ButtonProps;
+
+const BackButton = ({ ...buttonProps }: SourceEditButtonProps): JSX.Element => {
+  const { t } = useTranslation();
+  const history = useHistory();
+
+  const handleBackClick = () => {
+    history.goBack();
+  };
+
+  return (
+    <Button
+      {...buttonProps}
+      startIcon={<ArrowBack />}
+      onClick={handleBackClick}
+      disableRipple
+    >
+      <Typography>{t("back")}</Typography>
+    </Button>
+  );
+};
+
+export default BackButton;
