+++ app/src/features/Sources/CredentialEditButton.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 
 import { Button, ButtonProps, Typography } from "@material-ui/core";
+import { useTranslation } from "react-i18next";
 import { useParams } from "react-router-dom";
 
 import { useAppDispatch } from "app/store";
@@ -13,9 +14,10 @@ type SourceEditButtonProps = ButtonProps;
 const CredentialEditButton = ({
   ...buttonProps
 }: SourceEditButtonProps): JSX.Element => {
+  const { t } = useTranslation();
   const dispatch = useAppDispatch();
-
   const { sourceId: id } = useParams<{ sourceId: string }>();
+
   const { data: source } = useApiSourcesRetrieveQuery({ id }, { skip: !id });
   const handleSourceEdit = () => {
     if (source) dispatch(editCredential(source));
@@ -23,7 +25,7 @@ const CredentialEditButton = ({
 
   return (
     <Button {...buttonProps} onClick={handleSourceEdit}>
-      <Typography>Database settings</Typography>
+      <Typography>{t("databaseSettings")}</Typography>
     </Button>
   );
 };
